% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_and_plots.R
\name{as_graph}
\alias{as_graph}
\title{Coerce a variety of inputs to igraph.}
\usage{
as_graph(x)
}
\arguments{
\item{x}{an igraph object, a two-column edgelist matrix, or an adjacency matrix.}
}
\value{
an undirected igraph object.
}
\description{
A pure constructor (no logging).
}
\examples{
el <- matrix(c(1, 2, 2, 3, 3, 1), ncol = 2, byrow = TRUE)
as_graph(el)
}
