% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcda_grasp.R
\name{nce_score}
\alias{nce_score}
\title{Global NCE leader score (Eq. 14).}
\usage{
nce_score(g, leaders, verbose = FALSE)
}
\arguments{
\item{g}{an igraph object (undirected, simple).}

\item{leaders}{integer vector of 1-based leader vertex indices.}

\item{verbose}{logical; emit a cli trace.}
}
\value{
scalar NCE score H(C).
}
\description{
Global NCE leader score (Eq. 14).
}
\examples{
g <- igraph::make_graph("Zachary")
nce_score(g, leaders = c(1, 34))
}
