% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_and_plots.R
\name{plot_reactive_pk}
\alias{plot_reactive_pk}
\title{Plot the evolution of selection probabilities p_k across iterations.}
\usage{
plot_reactive_pk(res, ...)
}
\arguments{
\item{res}{an `lcda_gr_result` object.}

\item{...}{further arguments passed to [matplot()].}
}
\value{
invisibly, the p_k history matrix.
}
\description{
Plot the evolution of selection probabilities p_k across iterations.
}
